Rename matrix_eq.rs -> assert_matrix_eq.rs
