+++ src/macros/mod.rs
@@ -7,7 +7,7 @@ mod vector;
 mod matrix;
 
 #[macro_use]
-mod matrix_eq;
+mod assert_matrix_eq;
 
 mod comparison;
 
@@ -22,7 +22,7 @@ pub use self::comparison::{
     ElementwiseComparator
 };
 
-pub use self::matrix_eq::{
+pub use self::assert_matrix_eq::{
     elementwise_matrix_comparison,
     elementwise_vector_comparison,
 };
